Debugged the debugging code and modularized data parsing.
The B-engine packet header information extraction from VDIF header is now available as three inline functions that may be called on host or device.

Interpretation of 4 lower bits in an integer as a single 2bit complex-valued sample is available as an inline function that may be called on host or device.