[pipeline/core] Added a basic pipeline framework module; address Issue #11.
This consists of a Steps class that takes an ordered list of functions or
other Steps that represent abstract "steps" that must be done in order on
an iterable (or a generator). The output of each step is then passed directly
to the input of the next step.

An Args class is provided for easily handling passing of data and/or meta-data
between steps but is not a requirement of the framework.