Update to quantization level definitions.
Old code defined levels, in order from most negative to most positive as [11, 10, 00, 01], i.e. the msb is a sign bit, and the lsb indicates the absolute value exceeds the threshold. Old code commented out, but still visible in file.

New code defines levels, in order from most negative to most positive as [00, 01, 10, 11], as required for the VDIF packing.