[aphids] Added an APHIDS library for common functions.
This defines an APHIDS context structure that should be initialized at the
start of an APHIDS hashpipe thread, with aphids_init. This context can then
be used to do various things. Presently it provides only two functions,
aphids_get and aphids_set, which get/set values on a local redis server.