[aphids_log] Replaced the use of syslog with a new aphids_log module.
This uses a REDIS pub/sub model to send log information to the local redis-
server which can then be read by any client subscribing to the aphids.*
channels, where the sub-channel represents the log level. Presently the logs
do not contain a time-stamp but that will be added in the future.