[aphids] Added input/output threads that read/write to a VDIF file respectively.
Presently the read thread is limited to reading from a local "sample_in.vdif" and will
read it over and over again onto the data buffer. Similarly the write thread is limited to
writing to a local "sample_out.vdif" and will overwrite it repeatedly until the hashpipe
instance is killed.

Obviously this is not idea as we would like to specify the actual files to read from and
written to. Also, these threads should wait until a file is given, and either read it or
write it once and then wait again for a new file name. These features are coming soon!