[aphids_start.py] The APHIDS startup script now uses argparse to read the command line.
Use ./aphids_start.py --help to see the list of arguments. By default just running the script
will start all null-type threads as instance 0 with CPU affinities as 1, 2, and 3. All of these
options however can be overwritten. In particular the thread types for the input and output threads
may be set to "file" while will use the file-type threads from the last commit.