Implemented handshake in communication protocol.
Frames are sent / received using tx_frames() and rx_frames() which wrap the lower-level tx_frame / rx_frame methods to send / receive all available frames within a single call in the transmitter / receiver thread. The wrapper methods also implement a handshake which allows the receiver to know how many frames to expect or to know when transmission will end, and which allows the transmitter to know when the receiver is done processing a transmission.