[vdif_in_file_thread] Transitioned to the new FSM model.
This thread will now wait in an "idle" state for a file name to be given to it
over a redis thread. Once it receives this it will switch to a "read" state and will
read VDIF packet blocks from the specified file and write them into its output buffer
until it reaches EOF. It will then go back to the "idle" state and wait for the next
file name.

Similar to the null thread it has an "error" state which it goes to in the event of an
error and waits forever to be killed.