[vdif_out_file_thread] Transitioned to the new FSM model.
Again, similar to previous threads. One major change here however is an "idle" state
similar to the input file thread. This thread will now wait on an input queue (in the form
of a redis list) for an output file name. Once it receives that it will go into its "write" state
which will overwrite a certain amount of data forever until it is killed. This is obviously non-ideal
but presently there's no way for it to know when it reaches an "EOF". More thought should go into
this.