Do not use cudaIpcMemHandle
The CUDA IPC shared memory cannot be shared among different threads
that belong to the same process. Trying to open a cudaIpcMemHandle
within the same process that created it resutls in cudaErrorInvalidDevice.

Memory allocated on GPU can be accessed directly by all threads, so
only pointers to that memory are now shared between threads.