Bugfix incorrect number of input VDIF frames passed to vdif_to_beng
The number of valid VDIF frames to process is given in the input
databuf structure, and may not always be equal to the total number
of VDIF frames in a complete B-engine frame. If processing assumes
more frames than the actual number of valid frames, then old data
for previous iterations is mixed into the output data.

Another side-effect of this is that old B-engine counts are encountered
which means that the output indexing can result in illegal memory
accesses.