Parse input VDIF packets for old SBDE bitcode
Day1 (official first day) and Day0 (test day) of March2015 campaign
used swarm_2_comp_2015_Mar_08_2108.bof on the SDBE which caused an
offset between the payload data and header data in the VDIF packets
associated with each of the four 10GbE streams connecting to the
Mark6.

This is the first commit of changes to the APHIDS code to accommodate
this data format.