Update to 10/11 SWARM rate
Changed constants in sample_rates.h and vdif_out_databuf.h as appropriate
for the higher sample rate.

Reconfigured the allocation of buffer space on the GPU in
vdif_inout_gpu_thread.cu. At this data rate B-engine frames are processed
in groups of 66, which means that there is not enough memory to allocate
all needed buffers for both channels simultaneously. Instead, only
gpu_A_0 and gpu_A_1 buffers are created separately, and a single
shared gpu_B_common buffer is used for both channels. This required
reordering the processing pipeline so that processing of data from
one channel reaches the A buffer before processing data from the other
channel.