Update code for 8pac data
8 BoVDIF packets are now sneaked into each raw VDIF packet read from
disk. This commit adapts the code to unpack the data accordingly.

Data frame counter and packet serial numbers no longer attempt to
remain strictly valid, and are simply that of the first packet in
the 8pac multiplied by 8 and incremented by the offset of a given
packet within the 8pac.