Adopt new B-engine header format
B-engine header layout and content has changed since 10/11 rate.
Mainly the B-engine counter ('b') field has been replaced by two
numbers, one counting the seconds since start of epoch (a la VDIF)
and the other counting the number of FPGA clock cycles since the
start of second, both valid at the start of the B-frame to which
the packet belongs. Since these counters span a total of 53 bits
the layout of the B-engine header was changed to reduce the FID
field to occupy only 3 bits. See design_files repo for more details.

The change to vdif_in_net_thread.c calls the beng_reform_headers
method which rewrites the header information of (unpacked 8pac)
B-engine-over-VDIF packets in the old style. The FID and CID values
can be copied directly, but the old style requires a B-frame number
not supported in the new format. As a workaround beng_reform.c uses
the B-engine timestamp to count the number of B-frames since the
first packet to derive a new B-frame number. This means that each
run of the processor will start counting at B-frame 0.