Do roll-by-58 in vdif_to_beng
Applying this data reordering is done by explicitly computing a
snapshot number for each datapoint read from the input buffer. Due to
this the increment on idx_beng_data_out is no longer needed, and an
offset to the packet-wide value is computed for each snapshot.

The snapshot number also simplifies the calculation of
offset_snapshot_index_by used to do the a-to-d shift.