Compute timestamp offset
The VDIF timestamp is only accurate to within 8 microseconds (data
length of one packet), whereas the B-engine timestamp reports time
to within ~0.25ns. The current implementation packs output VDIF data
starting with the first sample in the first B-engine frame so that
there may be an up to 8 microseconds offset in the timestamp applied.
To account for this the offset is calculated and reported in the log
to stdout.