[sgcomm] Check external dependency libscatgat
Main targets are now dependent on the availability of libscatgat.so
in the /src directory. If the library is not found an error is printed
to stdout and compilation fails.

The required library can be compiled from source available in this
repository on github: https://github.com/anyoung/scatter-gather.git