[aphids & scripting] Add flexible VDIF metadata
Some VDIF header fields in the output data can now be controlled
using command-line arguments to ctrl_aphids.py. The user can
currently specify:
  * two-letter station code
  * SWARM quadrant
  * SWARM receiver sideband
  * SWARM frequency band
The quadrant, receiver sideband, and frequency band map to receiver
and BDC sideband flags in the VDIF header. The station code is
passed as-is.

Communication between ctrl_aphids.py and the aphids process is
through the redis server.